+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - better import error messages at parsing
 - format can now handle any value type
 - updated the tests to use the new standard library, and testing every VM instruction and builtins (we didn't test everything before, this way we can be sure we don't break *anything* in the VM after each update)
+- renaming builtins to add a namespace to them (math:, sys:, str:, list: and such)
 
 ### Removed
 - removed `fmt/format` from our dependencies